Moved filter.js to utils
